Refactor market actor middleware tests
