+++ source/IntegrationTests/Api/WhenMarketActorAuthenticatorMiddlewareIsCalledTests.cs
@@ -19,6 +19,7 @@ using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Threading.Tasks;
 using Dapper;
+using Energinet.DataHub.Core.App.FunctionApp;
 using Energinet.DataHub.EDI.Api.Configuration.Middleware.Authentication;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.DataAccess;
@@ -36,13 +37,13 @@ using Xunit.Categories;
 namespace Energinet.DataHub.EDI.IntegrationTests.Api;
 
 [IntegrationTest]
-public class WhenActorIsTryingToAuthenticateTests : TestBase
+public class WhenMarketActorAuthenticatorMiddlewareIsCalledTests : TestBase
 {
     private readonly NextSpy _nextSpy;
     private readonly FunctionExecutionDelegate _next;
     private readonly FunctionContextBuilder _functionContextBuilder;
 
-    public WhenActorIsTryingToAuthenticateTests(IntegrationTestFixture integrationTestFixture)
+    public WhenMarketActorAuthenticatorMiddlewareIsCalledTests(IntegrationTestFixture integrationTestFixture)
         : base(integrationTestFixture)
     {
         AuthenticatedActor.SetAuthenticatedActor(null);
@@ -199,6 +200,46 @@ public class WhenActorIsTryingToAuthenticateTests : TestBase
 
         var functionContext = _functionContextBuilder
             .TriggeredByHttp(withContentType: null, withToken: token)
+            .Build();
+
+        var sut = CreateMarketActorAuthenticatorMiddleware();
+
+        // Act
+        await sut.Invoke(functionContext, _next);
+
+        // Assert
+        Assert.True(_nextSpy.NextWasCalled);
+    }
+
+    [Theory]
+    [InlineData(TriggerType.TimerTrigger)]
+    [InlineData(TriggerType.ServiceBusTrigger)]
+    public async Task When_calling_authentication_middleware_without_being_a_http_trigger_then_next_is_called(TriggerType triggerType)
+    {
+        // Arrange
+        var functionContext = _functionContextBuilder
+            .WithTriggeredBy(triggerType)
+            .Build();
+
+        var sut = CreateMarketActorAuthenticatorMiddleware();
+
+        // Act
+        await sut.Invoke(functionContext, _next);
+
+        // Assert
+        Assert.True(_nextSpy.NextWasCalled);
+    }
+
+    [Theory]
+    [InlineData("HealthCheck")]
+    [InlineData("RenderSwaggerUI")]
+    [InlineData("RenderSwaggerDocument")]
+    public async Task When_calling_authentication_middleware_with_specific_function_name_then_next_is_called(string functionName)
+    {
+        // Arrange
+        var functionContext = _functionContextBuilder
+            .TriggeredByHttp()
+            .WithFunctionName(functionName)
             .Build();
 
         var sut = CreateMarketActorAuthenticatorMiddleware();
